+++ src/containers/subjects/IGV/utils.ts
@@ -1,5 +1,6 @@
 import config from '../../../config';
-import { getBaseNameFromKey, constructGDSUrl } from '../../../API/subject';
+import { getBaseNameFromKey } from '../../../api/utils';
+import { constructGDSUrl } from '../../../api/gds';
 import igv, { IGVBrowser, ITrack } from 'igv';
 import API from '@aws-amplify/api';
 import genomes from './genomes';
